+++ astrocoffee/__init__.py
@@ -1,5 +1,7 @@
+# -*- coding: utf-8 -*-
+
 __author__ = 'Anowar J. Shajib'
 __email__ = 'ajshajib@gmail.com'
-__version__ = '0.0.1'
+__version__ = '0.0.2'
 
 from .astrocoffee import app