+++ astrocoffee/astrocoffee.py
@@ -1,23 +1,30 @@
 # -*- coding: utf-8 -*-
 """
-    AstroCoffee
-    ~~~~~~
-    Web application for organizing astro-ph coffee discussion.
+AstroCoffee
+~~~~~~~~~~~
+Web application for organizing astro-ph coffee discussion.
 
-    :copyright: © 2018 by Anowar J. Shajib.
-    :license: MIT.
+Inspired by Astro Coffee 2 by Abhimat Gautam and previous versions of
+astroph.py by Ryan T. Hamilton, Ian J. Crossfield, and Nathaniel Ross.
+
+:author: Anowar J. Shajib and Abhimat Gautam.
+:copyright: © 2018 by Anowar J. Shajib.
+:license: MIT.
 """
 
 import os
 import sqlite3
 from datetime import datetime, timedelta
+import click
 from flask import Flask, request, session, g, redirect, url_for, abort, \
     render_template, flash, escape
+from flask_bcrypt import Bcrypt
 
 from .web import get_paper
 
 
 app = Flask(__name__, instance_relative_config=True)
+bcrypt = Bcrypt(app)
 
 app.config.from_object('config')
 app.config.from_pyfile('config.py')
@@ -29,7 +36,6 @@ app.config.update(dict(
 
 def connect_db():
     """Connect to the specific database."""
-
     rv = sqlite3.connect(app.config['DATABASE'])
     rv.row_factory = sqlite3.Row
     return rv
@@ -47,10 +53,67 @@ def init_db():
 
 @app.cli.command('initdb')
 def initdb_command():
-    """Initialize the database from command $flask initdb."""
+    """Initialize the database.
 
+    :Example:
+
+    .. code-block: bash
+
+        $ flask initdb
+
+    """
     init_db()
-    print('Initialized the AstroCoffee database.')
+    click.echo('Initialized the AstroCoffee database.')
+
+
+@app.cli.command('mkuser')
+@click.option('--user', prompt='Username', help='Login username.')
+@click.option('--password', prompt='Password', help='Login password.',
+              hide_input=True)
+@click.option('--repassword', prompt='Re-enter password',
+              help='Confirm login password.', hide_input=True)
+@click.option('--salt', prompt='Secret key', help='Phrase for salting.')
+def mkuser_command(user, password, repassword, salt):
+    """
+    Create the user with username and password.
+
+    :Example:
+
+    .. code-block:: bash
+
+        $ flask mkuser
+
+    :param user: Username.
+    :type user: str
+    :param password: Password.
+    :type password: str
+    :param repassword: Re-entered password.
+    :type repassword: str
+    :param salt: Phrase for salting.
+    :type salt: str
+    :return: None
+    :rtype: None
+    """
+    if password != repassword:
+        click.echo('Passwords do not match!')
+    else:
+        with open(os.path.join(app.root_path,
+                               '../instance/config.py'), 'w') as f:
+            salted_password = salt + password
+            hashed_password = bcrypt.generate_password_hash(salted_password)
+            f.write(('# -*- coding: utf-8 -*-\n'
+                     'SECRET_KEY = "{}"\n'
+                     'USERNAME = "{}"\n'
+                     'PASSWORD = {}').format(salt, user, hashed_password))
+
+
+@app.cli.command('check')
+@click.option('--p', prompt='Username', help='Login username.')
+def check(p):
+    salted_password = app.config['SECRET_KEY'] + p.encode('utf-8')
+    click.echo(salted_password)
+    click.echo(bcrypt.check_password_hash(app.config['PASSWORD'],
+                                          salted_password))
 
 
 def get_db():
@@ -71,7 +134,7 @@ def close_db(error):
 
 @app.template_filter('strftime')
 def _jinja2_filter_datetime(date, format_=None):
-    """Convert datetime to given format. Filter for Jinja2."""
+    """Convert datetime to given format."""
     date = datetime.strptime(date, '%Y-%m-%d %H:%M:%S')
     if format_ is None:
         format_ = '%b %d %Y'
@@ -88,9 +151,9 @@ def show_papers():
     week_earlier = week_earlier.strftime("%Y-%m-%d %H:%M")
 
     db = get_db()
-    cursor = db.execute(('select url, author, title, abstract, date_extended, '
-                         'sources, volunteer, discussed from paper '
-                         'where date_extended between '
+    cursor = db.execute(('select id, url, author, title, abstract, '
+                         'date_extended, sources, volunteer, discussed '
+                         'from paper where date_extended between '
                          '"{}" and "{}"'.format(week_earlier, now)))
     papers = cursor.fetchall()
 
@@ -108,7 +171,7 @@ def show_papers():
 
 @app.route('/submit', methods=['POST'])
 def submit_paper():
-    """Save the paper info in the database after a URL submission"""
+    """Save the paper info in the database after a URL submission."""
     paper = get_paper(escape(request.form['article']))
 
     if paper is not None:
@@ -134,6 +197,7 @@ def submit_paper():
 
     return redirect(url_for('show_papers'))
 
+
 @app.route('/archive/')
 def show_archive():
     """Render the archive page."""
@@ -142,5 +206,64 @@ def show_archive():
 
 @app.route('/usefullinks/')
 def show_useful_links():
-    """Render the useful links page"""
+    """Render the useful links page."""
     return render_template('usefullinks.html')
+
+
+@app.route('/login-form/')
+def show_login_form():
+    """Render the login page."""
+    return render_template('form.html')
+
+
+@app.route('/login', methods=['GET', 'POST'])
+def login():
+    """Log in the user."""
+    if request.method == 'POST':
+        salted_password = app.config['SECRET_KEY'] + request.form['password']
+
+        if request.form['username'] == app.config['USERNAME'] and \
+                bcrypt.check_password_hash(app.config['PASSWORD'],
+                                           salted_password):
+            session['logged_in'] = True
+            flash('You are logged in.', 'success')
+            return redirect(url_for('show_papers'))
+
+    flash('Invalid login credential(s).', 'error')
+    return render_template('form.html')
+
+
+@app.route('/logout')
+def logout():
+    """Log a user out."""
+    session.pop('logged_in', None)
+    flash('You are logged out.', 'success')
+    return redirect(url_for('show_papers'))
+
+
+@app.route('/mark_discussed')
+def mark_discussed():
+    """Mark a paper as discussed."""
+    paper_id = request.args.get('paper_id', None)
+
+    if paper_id is not None and session['logged_in']:
+        db = get_db()
+        db.execute('update paper set discussed = 1 where id = ?', [paper_id])
+        db.commit()
+        flash('The paper has been marked as discussed.', 'success')
+
+    return redirect(url_for('show_papers'))
+
+
+@app.route('/unmark_discussed')
+def unmark_discussed():
+    """Mark a paper as discussed."""
+    paper_id = request.args.get('paper_id', None)
+
+    if paper_id is not None and session['logged_in']:
+        db = get_db()
+        db.execute('update paper set discussed = 0 where id = ?', [paper_id])
+        db.commit()
+        flash('The paper has been marked as undiscussed.', 'success')
+
+    return redirect(url_for('show_papers'))