+++ HISTORY.rst
@@ -8,3 +8,10 @@ History
 
 * Working version with minimal features.
 * Supports only arXiv url/ID.
+
+0.0.2 (2018-03-28)
+++++++++++++++++++
+
+* Supports only arXiv url/ID.
+* Login feature to mark/unmark papers as discussed.
+* Flash message feedback on user interaction.